Extracted common logic to the AbstractEntityS3Service for potential reuse
