+++ ramls/bulkUpsertRequest.json
@@ -1,11 +1,11 @@
 {
   "$schema": "http://json-schema.org/draft-04/schema#",
-  "description": "Bulk instances upsert request",
+  "description": "Bulk entities upsert request",
   "type": "object",
   "additionalProperties": false,
   "properties": {
     "recordsFileName": {
-      "description": "File name of instances records",
+      "description": "File name of entities records",
       "type": "string"
     }
   },
