+++ src/main/java/org/folio/rest/impl/InstanceStorageInstancesBulkApi.java
@@ -5,23 +5,23 @@ import io.vertx.core.Context;
 import io.vertx.core.Handler;
 import java.util.Map;
 import javax.ws.rs.core.Response;
-import org.folio.rest.jaxrs.model.InstanceBulkRequest;
+import org.folio.rest.jaxrs.model.BulkUpsertRequest;
 import org.folio.rest.jaxrs.resource.InstanceStorageInstancesBulk;
 import org.folio.rest.support.EndpointFailureHandler;
-import org.folio.services.instance.InstanceS3Service;
+import org.folio.services.bulkprocessing.InstanceS3Service;
 import org.folio.services.s3storage.FolioS3ClientFactory;
 
 public class InstanceStorageInstancesBulkApi implements InstanceStorageInstancesBulk {
 
   @Override
-  public void postInstanceStorageInstancesBulk(InstanceBulkRequest bulkRequest, Map<String, String> okapiHeaders,
+  public void postInstanceStorageInstancesBulk(BulkUpsertRequest bulkRequest, Map<String, String> okapiHeaders,
                                                Handler<AsyncResult<Response>> asyncResultHandler,
                                                Context vertxContext) {
     new InstanceS3Service(new FolioS3ClientFactory(), vertxContext.owner(), okapiHeaders)
       .processBulkUpsert(bulkRequest)
       .map(PostInstanceStorageInstancesBulkResponse::respond201WithApplicationJson)
       .map(Response.class::cast)
-      .onFailure(EndpointFailureHandler::failureResponse)
+      .otherwise(EndpointFailureHandler::failureResponse)
       .onComplete(asyncResultHandler);
   }
